Renames files for clarity
