+++ le_utils/constants/embed_request.py
